+++ src/eve-server/manufacturing/RamProxyService.cpp
@@ -238,8 +238,8 @@ PyResult RamProxyService::Handle_InstallJob(PyCallArgs &call) {
             // consume required materials
             for(; curi != endi; curi++) {
                 if((*curi)->typeID() == cur->typeID && (*curi)->ownerID() == call.client->GetCharacterID()) {
-                    if(qtyNeeded >= (uint32)(*curi)->quantity()) {
-                        qtyNeeded -= (uint32)(*curi)->quantity();
+                    if(qtyNeeded >= (*curi)->quantity()) {
+                        qtyNeeded -= (*curi)->quantity();
                         (*curi)->Delete();
                     } else {
                         (*curi)->AlterQuantity(-(int32)qtyNeeded);
@@ -729,7 +729,7 @@ void RamProxyService::_VerifyInstallJob_Install(const Rsp_InstallJob &rsp, const
                 if(    (*curi)->typeID() == cur->typeID
                     && (*curi)->ownerID() == c->GetCharacterID()
                 ) {
-                    if((uint32)(*curi)->quantity() < qtyNeeded)
+                    if((*curi)->quantity() < qtyNeeded)
                         qtyNeeded -= (*curi)->quantity();
                     else
                         break;
